Moved includes to .h files, made the tetromino_type enum private, created an OccupyGridPos function, used the OccupyGridPos to deoccupy the tetrominoes current pos so it doesn't conflate its occupying of its position as someone else's, and then renamed the currPos variable in the draw function
I moved all includes to header files to be more consistent stylistically, making the tetromino_type enum private is because realistically no function has reason to use it. And an OccupyGridPos function to move break down existing functions, its usage in ChangePos is necessary to not allow the tetromino to think someone else is occupying that position when it is doing that itself.