Modified the endgame state, a drawing order, and properly checking for filled rows.
I included a logic function in the grid that manages to check for the endgame, and for any filled rows. Rows are drawn according to the drawing order which means we can adjust the placing of each row, this will be important in emptying out any filled rows and moving them to the top.