Fixed OutOfBounds spawning, added a printGrid function, removed unnecessary brackets to declutter
I fixed an OutOfBounds spawn bug, where the tetromino spanws out of bounds and the gameController tries to reset it multiple times and changes its color and type (see issue#6). I added a printGrid function, I only used it for debugging; it could be useful later on. I removed unnecessary brackets in one-line if statements, for loops, and while loops; this doesn't affect the program post-compile.